@@ -6,11 +6,16 @@ description: Demonstrates how to use the ar_flutter_plugin plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.16.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  permission_handler: 10.1.0
+  vector_math: 2.1.1
+  json_annotation: 4.5.0
+  geolocator: 9.0.0
+  app_settings: 4.1.6
 
   ar_flutter_plugin:
     # When depending on this package from a real application you should use:
