@@ -1,20 +1,21 @@
 name: ar_flutter_plugin
 description: Flutter Plugin for creating (collaborative) Augmented Reality experiences - Supports ARKit for iOS and ARCore for Android devices.
-version: 0.7.2
+version: 0.7.1
 homepage: https://lars.carius.io
 repository: https://github.com/CariusLars/ar_flutter_plugin
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: ">=2.16.1 <3.0.0"
+  #flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  permission_handler: ^10.0.0
-  vector_math: ^2.1.2
-  json_annotation: ^4.5.0
-  geolocator: ^9.0.0
+  permission_handler: 10.1.0
+  vector_math: 2.1.1
+  json_annotation: 4.5.0
+  geolocator: 9.0.0
+  app_settings: 4.1.6
 
 dev_dependencies:
   flutter_test:
