@@ -19,7 +19,7 @@ import 'package:geoflutterfire/geoflutterfire.dart';
 import 'package:geolocator/geolocator.dart';
 
 class ExternalModelManagementWidget extends StatefulWidget {
-  ExternalModelManagementWidget({Key key}) : super(key: key);
+  ExternalModelManagementWidget({Key? key}) : super(key: key);
   @override
   _ExternalModelManagementWidgetState createState() =>
       _ExternalModelManagementWidgetState();
@@ -33,10 +33,10 @@ class _ExternalModelManagementWidgetState
   FirebaseManager firebaseManager = FirebaseManager();
   Map<String, Map> anchorsInDownloadProgress = Map<String, Map>();
 
-  ARSessionManager arSessionManager;
-  ARObjectManager arObjectManager;
-  ARAnchorManager arAnchorManager;
-  ARLocationManager arLocationManager;
+  ARSessionManager? arSessionManager;
+  ARObjectManager? arObjectManager;
+  ARAnchorManager? arAnchorManager;
+  ARLocationManager? arLocationManager;
 
   List<ARNode> nodes = [];
   List<ARAnchor> anchors = [];
@@ -63,7 +63,7 @@ class _ExternalModelManagementWidgetState
   @override
   void dispose() {
     super.dispose();
-    arSessionManager.dispose();
+    arSessionManager!.dispose();
   }
 
   @override
@@ -165,22 +165,22 @@ class _ExternalModelManagementWidgetState
     this.arAnchorManager = arAnchorManager;
     this.arLocationManager = arLocationManager;
 
-    this.arSessionManager.onInitialize(
+    this.arSessionManager!.onInitialize(
           showFeaturePoints: false,
           showPlanes: true,
           customPlaneTexturePath: "Images/triangle.png",
           showWorldOrigin: true,
         );
-    this.arObjectManager.onInitialize();
-    this.arAnchorManager.initGoogleCloudAnchorMode();
+    this.arObjectManager!.onInitialize();
+    this.arAnchorManager!.initGoogleCloudAnchorMode();
 
-    this.arSessionManager.onPlaneOrPointTap = onPlaneOrPointTapped;
-    this.arObjectManager.onNodeTap = onNodeTapped;
-    this.arAnchorManager.onAnchorUploaded = onAnchorUploaded;
-    this.arAnchorManager.onAnchorDownloaded = onAnchorDownloaded;
+    this.arSessionManager!.onPlaneOrPointTap = onPlaneOrPointTapped;
+    this.arObjectManager!.onNodeTap = onNodeTapped;
+    this.arAnchorManager!.onAnchorUploaded = onAnchorUploaded;
+    this.arAnchorManager!.onAnchorDownloaded = onAnchorDownloaded;
 
     this
-        .arLocationManager
+        .arLocationManager!
         .startLocationUpdates()
         .then((value) => null)
         .onError((error, stackTrace) {
@@ -192,7 +192,7 @@ class _ExternalModelManagementWidgetState
                 "Action Required",
                 "To use cloud anchor functionality, please enable your location services",
                 "Settings",
-                this.arLocationManager.openLocationServicesSettings,
+                this.arLocationManager!.openLocationServicesSettings,
                 "Cancel");
             break;
           }
@@ -204,7 +204,7 @@ class _ExternalModelManagementWidgetState
                 "Action Required",
                 "To use cloud anchor functionality, please allow the app to access your device's location",
                 "Retry",
-                this.arLocationManager.startLocationUpdates,
+                this.arLocationManager!.startLocationUpdates,
                 "Cancel");
             break;
           }
@@ -216,24 +216,24 @@ class _ExternalModelManagementWidgetState
                 "Action Required",
                 "To use cloud anchor functionality, please allow the app to access your device's location",
                 "Settings",
-                this.arLocationManager.openAppPermissionSettings,
+                this.arLocationManager!.openAppPermissionSettings,
                 "Cancel");
             break;
           }
 
         default:
           {
-            this.arSessionManager.onError(error.toString());
+            this.arSessionManager!.onError(error.toString());
             break;
           }
       }
-      this.arSessionManager.onError(error.toString());
+      this.arSessionManager!.onError(error.toString());
     });
   }
 
   void onModelSelected(AvailableModel model) {
     this.selectedModel = model;
-    this.arSessionManager.onError(model.name + " selected");
+    this.arSessionManager!.onError(model.name + " selected");
     setState(() {
       modelChoiceActive = false;
     });
@@ -241,7 +241,7 @@ class _ExternalModelManagementWidgetState
 
   Future<void> onRemoveEverything() async {
     anchors.forEach((anchor) {
-      this.arAnchorManager.removeAnchor(anchor);
+      this.arAnchorManager!.removeAnchor(anchor);
     });
     anchors = [];
     if (lastUploadedAnchor != "") {
@@ -258,9 +258,8 @@ class _ExternalModelManagementWidgetState
   }
 
   Future<void> onNodeTapped(List<String> nodeNames) async {
-    var foregroundNode =
-        nodes.firstWhere((element) => element.name == nodeNames.first);
-    this.arSessionManager.onError(foregroundNode.data["onTapText"]);
+    var foregroundNode = nodes.firstWhere((element) => element.name == nodeNames.first);
+    this.arSessionManager!.onError(foregroundNode.data!["onTapText"]);
   }
 
   Future<void> onPlaneOrPointTapped(
@@ -270,8 +269,8 @@ class _ExternalModelManagementWidgetState
     if (singleHitTestResult != null) {
       var newAnchor = ARPlaneAnchor(
           transformation: singleHitTestResult.worldTransform, ttl: 2);
-      bool didAddAnchor = await this.arAnchorManager.addAnchor(newAnchor);
-      if (didAddAnchor) {
+      bool? didAddAnchor = await this.arAnchorManager!.addAnchor(newAnchor);
+      if (didAddAnchor!) {
         this.anchors.add(newAnchor);
         // Add note to anchor
         var newNode = ARNode(
@@ -281,24 +280,24 @@ class _ExternalModelManagementWidgetState
             position: VectorMath.Vector3(0.0, 0.0, 0.0),
             rotation: VectorMath.Vector4(1.0, 0.0, 0.0, 0.0),
             data: {"onTapText": "I am a " + this.selectedModel.name});
-        bool didAddNodeToAnchor =
-            await this.arObjectManager.addNode(newNode, planeAnchor: newAnchor);
-        if (didAddNodeToAnchor) {
+        bool? didAddNodeToAnchor =
+            await this.arObjectManager!.addNode(newNode, planeAnchor: newAnchor);
+        if (didAddNodeToAnchor!) {
           this.nodes.add(newNode);
           setState(() {
             readyToUpload = true;
           });
         } else {
-          this.arSessionManager.onError("Adding Node to Anchor failed");
+          this.arSessionManager!.onError("Adding Node to Anchor failed");
         }
       } else {
-        this.arSessionManager.onError("Adding Anchor failed");
+        this.arSessionManager!.onError("Adding Anchor failed");
       }
     }
   }
 
   Future<void> onUploadButtonPressed() async {
-    this.arAnchorManager.uploadAnchor(this.anchors.last);
+    this.arAnchorManager!.uploadAnchor(this.anchors.last);
     setState(() {
       readyToUpload = false;
     });
@@ -307,7 +306,7 @@ class _ExternalModelManagementWidgetState
   onAnchorUploaded(ARAnchor anchor) {
     // Upload anchor information to firebase
     firebaseManager.uploadAnchor(anchor,
-        currentLocation: this.arLocationManager.currentLocation);
+        currentLocation: this.arLocationManager!.currentLocation);
     // Upload child nodes to firebase
     if (anchor is ARPlaneAnchor) {
       anchor.childNodes.forEach((nodeName) => firebaseManager.uploadObject(
@@ -317,20 +316,19 @@ class _ExternalModelManagementWidgetState
       readyToDownload = true;
       readyToUpload = false;
     });
-    this.arSessionManager.onError("Upload successful");
+    this.arSessionManager!.onError("Upload successful");
   }
 
-  ARAnchor onAnchorDownloaded(Map<String, dynamic> serializedAnchor) {
-    final anchor = ARPlaneAnchor.fromJson(
-        anchorsInDownloadProgress[serializedAnchor["cloudanchorid"]]);
+  ARAnchor onAnchorDownloaded(Map<String,dynamic> serializedAnchor) {
+    final anchor = ARPlaneAnchor.fromJson(anchorsInDownloadProgress[serializedAnchor["cloudanchorid"]] as Map<String,dynamic>);
     anchorsInDownloadProgress.remove(anchor.cloudanchorid);
     this.anchors.add(anchor);
 
     // Download nodes attached to this anchor
     firebaseManager.getObjectsFromAnchor(anchor, (snapshot) {
       snapshot.docs.forEach((objectDoc) {
-        ARNode object = ARNode.fromMap(objectDoc.data());
-        arObjectManager.addNode(object, planeAnchor: anchor);
+        ARNode object = ARNode.fromMap(objectDoc.data() as Map<String, dynamic>);
+        arObjectManager!.addNode(object, planeAnchor: anchor);
         this.nodes.add(object);
       });
     });
@@ -347,18 +345,18 @@ class _ExternalModelManagementWidgetState
     //});
 
     // Get anchors within a radius of 100m of the current device's location
-    if (this.arLocationManager.currentLocation != null) {
+    if (this.arLocationManager!.currentLocation != null) {
       firebaseManager.downloadAnchorsByLocation((snapshot) {
         final cloudAnchorId = snapshot.get("cloudanchorid");
-        anchorsInDownloadProgress[cloudAnchorId] = snapshot.data();
-        arAnchorManager.downloadAnchor(cloudAnchorId);
-      }, this.arLocationManager.currentLocation, 0.1);
+        anchorsInDownloadProgress[cloudAnchorId] = snapshot.data() as Map<String, dynamic>;
+        arAnchorManager!.downloadAnchor(cloudAnchorId);
+      }, this.arLocationManager!.currentLocation, 0.1);
       setState(() {
         readyToDownload = false;
       });
     } else {
       this
-          .arSessionManager
+          .arSessionManager!
           .onError("Location updates not running, can't download anchors");
     }
   }
@@ -406,11 +404,11 @@ typedef FirebaseDocumentStreamListener = void Function(
     DocumentSnapshot snapshot);
 
 class FirebaseManager {
-  FirebaseFirestore firestore;
-  Geoflutterfire geo;
-  CollectionReference anchorCollection;
-  CollectionReference objectCollection;
-  CollectionReference modelCollection;
+  FirebaseFirestore? firestore;
+  Geoflutterfire? geo;
+  CollectionReference? anchorCollection;
+  CollectionReference? objectCollection;
+  CollectionReference? modelCollection;
 
   // Firebase initialization function
   Future<bool> initializeFlutterFire() async {
@@ -428,7 +426,7 @@ class FirebaseManager {
     }
   }
 
-  void uploadAnchor(ARAnchor anchor, {Position currentLocation}) {
+  void uploadAnchor(ARAnchor anchor, {Position? currentLocation}) {
     if (firestore == null) return;
 
     var serializedAnchor = anchor.toJson();
@@ -437,13 +435,13 @@ class FirebaseManager {
     serializedAnchor["expirationTime"] = expirationTime;
     // Add location
     if (currentLocation != null) {
-      GeoFirePoint myLocation = geo.point(
+      GeoFirePoint myLocation = geo!.point(
           latitude: currentLocation.latitude,
           longitude: currentLocation.longitude);
       serializedAnchor["position"] = myLocation.data;
     }
 
-    anchorCollection
+    anchorCollection!
         .add(serializedAnchor)
         .then((value) =>
             print("Successfully added anchor: " + serializedAnchor["name"]))
@@ -455,7 +453,7 @@ class FirebaseManager {
 
     var serializedNode = node.toMap();
 
-    objectCollection
+    objectCollection!
         .add(serializedNode)
         .then((value) =>
             print("Successfully added object: " + serializedNode["name"]))
@@ -463,7 +461,7 @@ class FirebaseManager {
   }
 
   void downloadLatestAnchor(FirebaseListener listener) {
-    anchorCollection
+    anchorCollection!
         .orderBy("expirationTime", descending: false)
         .limitToLast(1)
         .get()
@@ -475,10 +473,10 @@ class FirebaseManager {
   void downloadAnchorsByLocation(FirebaseDocumentStreamListener listener,
       Position location, double radius) {
     GeoFirePoint center =
-        geo.point(latitude: location.latitude, longitude: location.longitude);
+        geo!.point(latitude: location.latitude, longitude: location.longitude);
 
-    Stream<List<DocumentSnapshot>> stream = geo
-        .collection(collectionRef: anchorCollection)
+    Stream<List<DocumentSnapshot>> stream = geo!
+        .collection(collectionRef: anchorCollection!)
         .within(center: center, radius: radius, field: 'position');
 
     stream.listen((List<DocumentSnapshot> documentList) {
@@ -491,7 +489,7 @@ class FirebaseManager {
   void downloadAnchorsByChannel() {}
 
   void getObjectsFromAnchor(ARPlaneAnchor anchor, FirebaseListener listener) {
-    objectCollection
+    objectCollection!
         .where("name", whereIn: anchor.childNodes)
         .get()
         .then((value) => listener(value))
@@ -500,13 +498,13 @@ class FirebaseManager {
 
   void deleteExpiredDatabaseEntries() {
     WriteBatch batch = FirebaseFirestore.instance.batch();
-    anchorCollection
+    anchorCollection!
         .where("expirationTime",
             isLessThan: DateTime.now().millisecondsSinceEpoch / 1000)
         .get()
         .then((anchorSnapshot) => anchorSnapshot.docs.forEach((anchorDoc) {
               // Delete all objects attached to the expired anchor
-              objectCollection
+              objectCollection!
                   .where("name", arrayContainsAny: anchorDoc.get("childNodes"))
                   .get()
                   .then((objectSnapshot) => objectSnapshot.docs.forEach(
@@ -518,7 +516,7 @@ class FirebaseManager {
   }
 
   void downloadAvailableModels(FirebaseListener listener) {
-    modelCollection
+    modelCollection!
         .get()
         .then((value) => listener(value))
         .catchError((error) => print("Failed to download objects: $error"));
@@ -536,7 +534,7 @@ class ModelSelectionWidget extends StatefulWidget {
   final Function onTap;
   final FirebaseManager firebaseManager;
 
-  ModelSelectionWidget({this.onTap, this.firebaseManager});
+  ModelSelectionWidget({required this.onTap, required this.firebaseManager});
 
   @override
   _ModelSelectionWidgetState createState() => _ModelSelectionWidgetState();
@@ -545,7 +543,7 @@ class ModelSelectionWidget extends StatefulWidget {
 class _ModelSelectionWidgetState extends State<ModelSelectionWidget> {
   List<AvailableModel> models = [];
 
-  String selected;
+  String? selected;
 
   @override
   void initState() {
